% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis-package.R
\docType{package}
\name{metis-package}
\alias{metis}
\alias{metis-package}
\title{metis: MICOM measurement invariance and multigroup analysis for seminr}
\description{
Implements the Measurement Invariance of Composite Models (MICOM) procedure
for \pkg{seminr} partial least squares structural equation models, with thin
wrappers for bootstrap and permutation multigroup analysis. Composites may use
correlation weights (reflective-style) or regression weights (formative).
Results have been cross-verified against \code{cSEM::testMICOM}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aar0ndaniel/metis}
  \item Report bugs at \url{https://github.com/aar0ndaniel/metis/issues}
}

}
\author{
\strong{Maintainer}: Aaron Akuteye \email{aaronakuteye@gmail.com}

}
\keyword{internal}
