% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mga.R
\name{print.metis_mga}
\alias{print.metis_mga}
\title{Print a multigroup analysis result}
\usage{
\method{print}{metis_mga}(x, ...)
}
\arguments{
\item{x}{An object of class \code{metis_mga}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a multigroup analysis result
}
