% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micom.R
\name{print.metis_micom}
\alias{print.metis_micom}
\title{Print a MICOM result}
\usage{
\method{print}{metis_micom}(x, ...)
}
\arguments{
\item{x}{An object of class \code{metis_micom}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a MICOM result
}
